darkdawn_pay/requirements/dev.txt
-r base.txt
ipython
black==22.3.0
django-debug-toolbar==3.2.2
docker==5.0.3
factory_boy==2.12.0
pdbpp==0.10.3
pytest-cov==3.0.0
pytest-django==4.4.0
pytest==6.2.5
pytest-sugar
requests-mock
isort[requirements_deprecated_finder]
